Clear bad model picker preference state
