+++ src/vs/workbench/test/browser/workbenchTestServices.ts
@@ -16,7 +16,6 @@ import { isValidBasename } from '../../../base/common/extpath.js';
 import { IMarkdownString } from '../../../base/common/htmlContent.js';
 import { Disposable, DisposableStore, IDisposable } from '../../../base/common/lifecycle.js';
 import { Schemas } from '../../../base/common/network.js';
-import { observableValue } from '../../../base/common/observable.js';
 import { posix, win32 } from '../../../base/common/path.js';
 import { IProcessEnvironment, isWindows, OperatingSystem } from '../../../base/common/platform.js';
 import { env } from '../../../base/common/process.js';
@@ -139,7 +138,7 @@ import { TerminalEditorInput } from '../../contrib/terminal/browser/terminalEdit
 import { IEnvironmentVariableService } from '../../contrib/terminal/common/environmentVariable.js';
 import { EnvironmentVariableService } from '../../contrib/terminal/common/environmentVariableService.js';
 import { IRegisterContributedProfileArgs, IShellLaunchConfigResolveOptions, ITerminalProfileProvider, ITerminalProfileResolverService, ITerminalProfileService, type ITerminalConfiguration } from '../../contrib/terminal/common/terminal.js';
-import { ChatEntitlement, IChatEntitlementService } from '../../services/chat/common/chatEntitlementService.js';
+import { IChatEntitlementService } from '../../services/chat/common/chatEntitlementService.js';
 import { IDecoration, IDecorationData, IDecorationsProvider, IDecorationsService, IResourceDecorationChangeEvent } from '../../services/decorations/common/decorations.js';
 import { CodeEditorService } from '../../services/editor/browser/codeEditorService.js';
 import { EditorPaneService } from '../../services/editor/browser/editorPaneService.js';
@@ -185,7 +184,7 @@ import { InMemoryWorkingCopyBackupService } from '../../services/workingCopy/com
 import { IWorkingCopyEditorService, WorkingCopyEditorService } from '../../services/workingCopy/common/workingCopyEditorService.js';
 import { IWorkingCopyFileService, WorkingCopyFileService } from '../../services/workingCopy/common/workingCopyFileService.js';
 import { IWorkingCopyService, WorkingCopyService } from '../../services/workingCopy/common/workingCopyService.js';
-import { TestContextService, TestExtensionService, TestFileService, TestHistoryService, TestLoggerService, TestMarkerService, TestProductService, TestStorageService, TestTextResourcePropertiesService, TestWorkspaceTrustManagementService, TestWorkspaceTrustRequestService } from '../common/workbenchTestServices.js';
+import { TestChatEntitlementService, TestContextService, TestExtensionService, TestFileService, TestHistoryService, TestLoggerService, TestMarkerService, TestProductService, TestStorageService, TestTextResourcePropertiesService, TestWorkspaceTrustManagementService, TestWorkspaceTrustRequestService } from '../common/workbenchTestServices.js';
 
 // Backcompat export
 export { TestFileService };
@@ -2172,35 +2171,6 @@ export async function workbenchTeardown(instantiationService: IInstantiationServ
 			editorGroupService.removeGroup(group);
 		}
 	});
-}
-
-export class TestChatEntitlementService implements IChatEntitlementService {
-
-	_serviceBrand: undefined;
-
-	readonly organisations: undefined;
-	readonly isInternal = false;
-	readonly sku = undefined;
-
-	readonly onDidChangeQuotaExceeded = Event.None;
-	readonly onDidChangeQuotaRemaining = Event.None;
-	readonly quotas = {};
-
-	update(token: CancellationToken): Promise<void> {
-		throw new Error('Method not implemented.');
-	}
-
-	readonly onDidChangeSentiment = Event.None;
-	readonly sentimentObs = observableValue({}, {});
-	readonly sentiment = {};
-
-	readonly onDidChangeEntitlement = Event.None;
-	entitlement: ChatEntitlement = ChatEntitlement.Unknown;
-	readonly entitlementObs = observableValue({}, ChatEntitlement.Unknown);
-
-	readonly anonymous = false;
-	onDidChangeAnonymous = Event.None;
-	readonly anonymousObs = observableValue({}, false);
 }
 
 export class TestContextMenuService implements IContextMenuService {
